GCI-332: Reorganise tests for clarity.
